+++ clang/include/clang/Lex/HeaderSearch.h
@@ -276,6 +276,15 @@ class HeaderSearch {
   /// a system header.
   std::vector<std::pair<std::string, bool>> SystemHeaderPrefixes;
 
+  /// External directories are user specified directories that are to be treated
+  /// like system directories for the purposes of warning suppression. A header
+  /// file that has a path that matches one of these prefixes is promoted to a
+  /// system header regardless of which header search path was used to resolve
+  /// the \#include directive. llvm::sys::path::remove_dots() is used to
+  /// normalize these paths by removing "." and ".." path components and
+  /// duplicate path separators. Trailing path separators are retained.
+  llvm::StringSet<llvm::BumpPtrAllocator> ExternalDirectoryPrefixes;
+
   /// The context hash used in SpecificModuleCachePath (unless suppressed).
   std::string ContextHash;
 
@@ -400,6 +409,9 @@ public:
     SearchDirs.push_back(dir);
     SearchDirsUsage.push_back(false);
   }
+
+  /// Add an additional external directory prefix path.
+  bool AddExternalDirectoryPrefix(StringRef Path);
 
   /// Set the list of system header prefixes.
   void SetSystemHeaderPrefixes(ArrayRef<std::pair<std::string, bool>> P) {